requirements.txt
alpaca-py==0.19.0
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.7.2
asttokens==2.4.1
black==24.3.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.2
coverage==7.4.4
debugpy==1.8.1
decorator==5.1.1
Deprecated==1.2.14
dnspython==2.6.1
duckdb==0.10.0
email_validator==2.1.1
executing==2.0.1
fastapi==0.110.0
googleapis-common-protos==1.63.0
grpcio==1.62.1
h11==0.14.0
httpcore==1.0.4
httptools==0.6.1
httpx==0.27.0
idna==3.6
importlib-metadata==6.11.0
iniconfig==2.0.0
ipykernel==6.29.3
ipython==8.22.2
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
jupyter_client==8.6.1
jupyter_core==5.7.2
loguru==0.7.2
MarkupSafe==2.1.5
matplotlib-inline==0.1.6
msgpack==1.0.8
mypy-extensions==1.0.0
nest-asyncio==1.6.0
numpy==1.26.4
opentelemetry-api==1.23.0
opentelemetry-distro==0.44b0
opentelemetry-exporter-otlp==1.23.0
opentelemetry-exporter-otlp-proto-common==1.23.0
opentelemetry-exporter-otlp-proto-grpc==1.23.0
opentelemetry-exporter-otlp-proto-http==1.23.0
opentelemetry-instrumentation==0.44b0
opentelemetry-instrumentation-asgi==0.44b0
opentelemetry-instrumentation-asyncio==0.44b0
opentelemetry-instrumentation-aws-lambda==0.44b0
opentelemetry-instrumentation-dbapi==0.44b0
opentelemetry-instrumentation-fastapi==0.44b0
opentelemetry-instrumentation-grpc==0.44b0
opentelemetry-instrumentation-httpx==0.44b0
opentelemetry-instrumentation-jinja2==0.44b0
opentelemetry-instrumentation-logging==0.44b0
opentelemetry-instrumentation-requests==0.44b0
opentelemetry-instrumentation-sqlite3==0.44b0
opentelemetry-instrumentation-system-metrics==0.44b0
opentelemetry-instrumentation-tornado==0.44b0
opentelemetry-instrumentation-tortoiseorm==0.44b0
opentelemetry-instrumentation-urllib==0.44b0
opentelemetry-instrumentation-urllib3==0.44b0
opentelemetry-instrumentation-wsgi==0.44b0
opentelemetry-propagator-aws-xray==1.0.1
opentelemetry-proto==1.23.0
opentelemetry-sdk==1.23.0
opentelemetry-semantic-conventions==0.44b0
opentelemetry-test-utils==0.44b0
opentelemetry-util-http==0.44b0
orjson==3.9.15
packaging==24.0
pandas==2.2.1
parso==0.8.3
pathspec==0.12.1
pexpect==4.9.0
platformdirs==4.2.0
pluggy==1.4.0
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==2.6.4
pydantic-extra-types==2.6.0
pydantic-settings==2.2.1
pydantic_core==2.16.3
Pygments==2.17.2
pytest==8.1.1
pytest-cov==4.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
pyzmq==25.1.2
requests==2.31.0
setuptools==69.2.0
six==1.16.0
sniffio==1.3.1
sseclient-py==1.8.0
stack-data==0.6.3
starlette==0.36.3
tornado==6.4
traitlets==5.14.2
typing_extensions==4.10.0
tzdata==2024.1
ujson==5.9.0
urllib3==2.2.1
uvicorn==0.28.0
uvloop==0.19.0
watchfiles==0.21.0
wcwidth==0.2.13
websockets==11.0.3
wrapt==1.16.0
zipp==3.18.1
